+++ CHANGELOG.md
@@ -2,9 +2,7 @@
 
 See README.md before updating this file.
 
-<!-- ## Unreleased [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.1.4...master) -->
-
-## v1.1.3 [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.1.3...v1.1.4)
+<!-- ## Unreleased [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.1.3...master) -->
 * Liquid Body Tag can Handle Liquid Templates [#64]
 * Add denormalized views over traffic tables. [#65]