+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 See README.md before updating this file.
 
-
 <!-- ## Unreleased [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.1.3...master) -->
-* Add denormalized views over traffic tables.
 * Add error logging for NewRelic when @tracker.track or @tracker.save are rescued
+* Liquid Body Tag can Handle Liquid Templates [#64]
+* Add denormalized views over traffic tables. [#65]
 
 ## v1.1.2 [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.1.2...v1.1.3)